change U64 definition to uint64_t and todo list

+++ src/bitboard_utils.h
@@ -1,4 +1,5 @@
-typedef unsigned long long U64;
+#include <stdint.h>
+typedef uint64_t U64;
 
 const U64 notAFile = 0xfefefefefefefefe; // ~0x0101010101010101
 const U64 notHFile = 0x7f7f7f7f7f7f7f7f; // ~0x8080808080808080
